+++ src/Models/Scopes/Product/WithProductAttributesScope.php
@@ -25,6 +25,7 @@ class WithProductAttributesScope implements Scope
         $builder
             ->addSelect($builder->getQuery()->from.'.entity_id AS id')
             ->addSelect($builder->getQuery()->from.'.sku')
+            ->addSelect($builder->getQuery()->from.'.visibility')
             ->addSelect($builder->getQuery()->from.'.type_id AS type');
 
         foreach ($attributes as $attribute) {